requirements.txt
streamlit>=1.32
plotly>=5.18
pandas>=2.2
pydantic>=2.6
pyyaml>=6.0.1
pyserial>=3.5
pymodbus>=3.6
watchdog>=3.0.0
openpyxl>=3.1
